+++ packages/vue/src/components/SunburstChart.ts
@@ -75,6 +75,7 @@ export const SunburstChart = defineComponent({
       wrapperClasses
     } = useChartInteraction<SunburstChartDatum>({
       hoverable: computed(() => props.hoverable),
+      showTooltip: computed(() => props.showTooltip),
       hoveredIndexProp: () => props.hoveredIndex,
       selectable: computed(() => props.selectable),
       selectedIndexProp: () => props.selectedIndex,
@@ -243,15 +244,14 @@ export const SunburstChart = defineComponent({
         }
       )
 
-      const tooltip =
-        props.showTooltip && props.hoverable
-          ? h(ChartTooltip, {
-              content: tooltipContent.value,
-              open: resolvedHoveredIndex.value !== null && tooltipContent.value !== '',
-              x: tooltipPosition.value.x,
-              y: tooltipPosition.value.y
-            })
-          : null
+      const tooltip = props.showTooltip
+        ? h(ChartTooltip, {
+            content: tooltipContent.value,
+            open: resolvedHoveredIndex.value !== null && tooltipContent.value !== '',
+            x: tooltipPosition.value.x,
+            y: tooltipPosition.value.y
+          })
+        : null
 
       if (!props.showLegend) {
         return h('div', { class: 'inline-block relative' }, [chart, tooltip])